Renamed filename to decimation_scheme
For consistency
