+++ experiment_prototype/decimation_scheme/decimation_scheme.py
